+++ plugins/09_Shadowlands/localization/ruRU.lua
@@ -370,6 +370,12 @@ L["addius_note"] = nil
 L["eyegor_note"] = nil
 L["sylla_note"] = nil
 
+L["grand_inquisitor_note"] = nil
+L["high_inquisitor_note"] = nil
+L["inquisitor_note"] = nil
+L["options_icons_inquisitors"] = nil
+L["options_icons_inquisitors_desc"] = nil
+
 -------------------------------------------------------------------------------
 ----------------------------------- THE MAW -----------------------------------
 -------------------------------------------------------------------------------